+++ src/main/java/com/ertho/gestiondestosck/services/impl/CommandeClientServiceImpl.java
@@ -153,6 +153,11 @@ public class CommandeClientServiceImpl implements CommandeClientService {
             log.error("Commande client ID is null");
             return;
         }
+        List<LigneCommandeClient> ligneCommandeClients = ligneCommandeClientRepository.findAllByCommandeClientId(id);
+        if(!ligneCommandeClients.isEmpty()){
+            throw new InvalidOperationException("Impossible de supprimer une commande client en cours ",
+                    ErrorCodes.COMMANDE_CLIENT_ALREADY_IN_USE);
+        }
         commandeClientRepository.deleteById(id);
     }
