+++ src/main/java/com/ertho/gestiondestosck/services/impl/CommandeFournisseurServiceImpl.java
@@ -243,12 +243,16 @@ public class CommandeFournisseurServiceImpl implements CommandeFournisseurServic
     }
 
 
-
     @Override
     public void delete(Integer id) {
         if(id == null){
             log.error("Commande fournisseur ID is null");
             return;
+        }
+        List<LigneCommandeFournisseur> ligneCommandeFournisseurs = ligneCommandeFournisseurRepository.findAllByCommandeFournisseurId(id);
+        if(!ligneCommandeFournisseurs.isEmpty()){
+            throw new InvalidOperationException("Impossible de supprimer une commande fournisseur en cours ",
+                    ErrorCodes.COMMANDE_FOURNISSEUR_ALREADY_IN_USE);
         }
         commandeFournisseurRepository.deleteById(id);
     }
